pkg/endpoint: move endpoint IP functions to ip.go
Signed-off by: Ian Vernon <ian@cilium.io>
