+++ pkg/endpoint/ip.go
@@ -1,0 +1,39 @@
+// Copyright 2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package endpoint
+
+import "net"
+
+// GetIPv4Address returns the IPv4 address of the endpoint
+func (e *Endpoint) GetIPv4Address() string {
+	return e.IPv4.String()
+}
+
+// GetIPv6Address returns the IPv6 address of the endpoint
+func (e *Endpoint) GetIPv6Address() string {
+	return e.IPv6.String()
+}
+
+// IPs returns the slice of valid IPs for this endpoint.
+func (e *Endpoint) IPs() []net.IP {
+	ips := []net.IP{}
+	if e.IPv4 != nil {
+		ips = append(ips, e.IPv4.IP())
+	}
+	if e.IPv6 != nil {
+		ips = append(ips, e.IPv6.IP())
+	}
+	return ips
+}
